requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
cloudinary
python-dotenv